Tweaked cleanup-backups.crt logging.
For some reason, it didn't log today's cleanup attempt, or the attempt didn't run, which would be odd. I tried a more sophisticated, cleaner logging style: redirecting stdout and stderr to file for the entire shell. We'll see how that goes. On the bright side, the backup script appears to be working flawlessly now.